Test Config extending.

+++ tests/Rah/Danpu/Test/BasicTest.php
@@ -6,6 +6,14 @@ use Rah\Danpu\Dump;
 use Rah\Danpu\Export;
 use Rah\Danpu\Import;
 
+class Config extends \Rah\Danpu\Config
+{
+    public $dsn = \test_db_dsn;
+    public $user = \test_db_user;
+    public $pass = \test_db_pass;
+    public $tmp = \test_tmp_dir;
+}
+
 class BasicTest extends \PHPUnit_Framework_TestCase
 {
     private $dump;
@@ -43,6 +51,16 @@ class BasicTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($files[0], $files[1]);
     }
 
+    /**
+     * @dataProvider provider
+     */
+
+    public function testExtending($source, $target)
+    {
+        $this->dump = new Dump(new Config);
+        $this->testDump($source, $target);
+    }
+
     public function provider()
     {
         $path = dirname(dirname(dirname(__DIR__))) . '/fixtures/*[.sql|.gz]';
@@ -73,6 +91,10 @@ class BasicTest extends \PHPUnit_Framework_TestCase
 
     public function tearDown()
     {
-        unlink($this->target);
+        if ($this->target)
+        {
+            unlink($this->target);
+            $this->target = null;
+        }
     }
 }
